WMM-6 composer support

+++ reset.sh
@@ -15,25 +15,26 @@ SCRIPT_DIR=$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)
 readonly SCRIPT_NAME SCRIPT_DIR
 #}}}
 
-# Accept envronment variable, or fallback to the script's directory
+# Accept environment variable, or fallback to the script's directory
 WMM_HOSTNAME=${WMM_HOSTNAME:-$(basename $PWD)}
-read -p "Purge the current VM $WMM_HOSTNAME? " -n 1 -r
-echo    # (optional) move to a new line
-if [[ $REPLY =~ ^[Yy]$ ]]
+
+if multipass list --format csv | grep -q "$WMM_HOSTNAME,"
 then
-	echo "Delete and purge the VM..."
-	multipass delete $WMM_HOSTNAME; multipass purge
+  read -p "Purge the existing VM $WMM_HOSTNAME? (y/N) " -n 1 -r
+  echo    # (optional) move to a new line
+  if [[ $REPLY =~ ^[Yy]$ ]]
+  then
+    echo "Delete and purge the VM..."
+    multipass delete $WMM_HOSTNAME; multipass purge
+  fi
 fi
 
-
-read -p "Replace the scripts and launch? " -n 1 -r
+read -p "Replace enabled scripts and launch? (y/N) " -n 1 -r
 echo    # (optional) move to a new line
 if [[ $REPLY =~ ^[Yy]$ ]]
 then
 	# Replace all the scripts
-	rm scripts-enabled/*.sh
-	cp scripts-available/*.sh scripts-enabled/
-	./host.sh
+	rm -rf scripts-enabled/*
+	cp -r scripts-available/* scripts-enabled/
+  ./host.sh
 fi
-
-
